Implement better bresenham procedure
There is a bug where infinite diagonal lines occasionally extend
from the cube. This appears to happen when dx and dy are both
equal to 1 in the bresenham procedure.

Interestingly, attempting to hard-code a particular line and
draw it on each frame does not result in the same drawing each
frame. On the first frame, an infinite line is drawn with the
intended slope. After the first frame, an infinite line is drawn
with a completely different slope. I discovered this by defining
start and end vectors in main.asm and passing them to the
bresenham procedure at the end of the logic for each frame.

If the program is modified so that those vectors are defined at a
different location in the data segment, the resulting slope
changes too. So there is probably some memory corruption going on.